feat: clean up and reuse stuff from dashboard
