+++ frontend/src/dashboard/service_provider/SpApplicationsTable.tsx
@@ -1,14 +1,14 @@
 // frontend/src/dashboard/SpApplicationsTable.tsx
 import { useNavigate } from "react-router-dom";
-import { useTranslateEnum } from "../intl/intl";
-import { EnumLabel } from "../intl/enum-labels";
-import { SimpleTable, Column } from "../components/SimpleTable";
-import { Badge } from "../components/ui";
-import { ENUM_KEY_PREFIX, getStatusVariant } from "./dashboardTableUtils";
-import { SpDashboardItem } from "./useSpDashboard";
+import { useTranslateEnum } from "../../intl/intl";
+import { EnumLabel } from "../../intl/enum-labels";
+import { SimpleTable, Column } from "../../components/SimpleTable";
+import { Badge } from "../../components/ui";
+import { ENUM_KEY_PREFIX, getStatusVariant } from "../shared/dashboardTableUtils";
+import { DashboardItem } from "../hooks/useDashboardApplications";
 
 type Props = {
-  items: SpDashboardItem[];
+  items: DashboardItem[];
   empty?: string;
 };
 
@@ -19,7 +19,7 @@ export const SpApplicationsTable = ({
   const navigate = useNavigate();
   const translateEnum = useTranslateEnum();
 
-  const columns: Column<SpDashboardItem>[] = [
+  const columns: Column<DashboardItem>[] = [
     {
       key: "typeLabel",
       header: "Type",
@@ -35,7 +35,7 @@ export const SpApplicationsTable = ({
       ),
     },
     {
-      key: "systemOperator",
+      key: "participant",
       header: <span className="hidden sm:inline">System Operator</span>,
       render: (value) => (
         <span className="hidden sm:inline">{String(value ?? "")}</span>
