+++ frontend/src/dashboard/shared/StatCard.tsx
@@ -1,0 +1,36 @@
+import { type ReactNode } from "react";
+import { Card, CardContent } from "../../components/ui";
+
+type StatCardProps = {
+  label: string;
+  value: number;
+  icon: ReactNode;
+  borderClass: string;
+  iconBgClass: string;
+};
+
+export const StatCard = ({
+  label,
+  value,
+  icon,
+  borderClass,
+  iconBgClass,
+}: StatCardProps) => (
+  <Card className={`flex-1 border-l-4 ${borderClass}`}>
+    <CardContent className="flex items-center gap-4 py-4">
+      <div
+        className={`w-10 h-10 rounded-lg flex items-center justify-center flex-shrink-0 ${iconBgClass}`}
+      >
+        {icon}
+      </div>
+      <div>
+        <p className="text-xs font-semibold uppercase tracking-wide text-semantic-text-subtle mb-1">
+          {label}
+        </p>
+        <p className="text-2xl font-bold text-semantic-text leading-none">
+          {value}
+        </p>
+      </div>
+    </CardContent>
+  </Card>
+);
