+++ frontend/src/dashboard/system_operator/SoStatCards.tsx
@@ -1,0 +1,40 @@
+import { IconClockCircle, IconWarningTriangle } from "@elhub/ds-icons";
+import { useDashboardApplications } from "../hooks/useDashboardApplications";
+import { StatCard } from "../shared/StatCard";
+import { Alert, Loader } from "@elhub/ds-components";
+
+export const SoStatCards = () => {
+  const { activeItems, isLoading, error } = useDashboardApplications();
+
+  if (isLoading) return <Loader size="small" />;
+  if (error) return <Alert variant="error">Failed to load stats.</Alert>;
+
+  return (
+    <div className="flex flex-col sm:flex-row gap-4">
+      <StatCard
+        label="Pending Applications"
+        value={activeItems.length}
+        icon={
+          <IconClockCircle
+            size="medium"
+            className="text-semantic-text-information"
+          />
+        }
+        borderClass="border-l-semantic-border-information"
+        iconBgClass="bg-semantic-background-information"
+      />
+      <StatCard
+        label="Resource Inconsistencies"
+        value={3}
+        icon={
+          <IconWarningTriangle
+            size="medium"
+            className="text-semantic-text-error"
+          />
+        }
+        borderClass="border-l-semantic-border-error"
+        iconBgClass="bg-semantic-background-error"
+      />
+    </div>
+  );
+};
